Updated root checker
The root checker will now allow the script run regardless of user. If sudo is not installed it will echo a warning and exit.  If run as root all commands will worked as expected. If run with sudo the command will work as expected.  If run as non-root without sudo the script will run as expected and prompt for the sudo password if needed depending on the selection option.